Move //types classes to subfolders These are getting quite large, and will be larger with the Huffman branch.
