+++ fly/parser/exceptions.cpp
@@ -1,7 +1,7 @@
 #include "fly/parser/exceptions.h"
 
 #include "fly/logger/logger.h"
-#include "fly/types/string.h"
+#include "fly/types/string/string.h"
 
 #include <cctype>
