+++ fly/parser/json_parser.cpp
@@ -1,7 +1,7 @@
 #include "fly/parser/json_parser.h"
 
 #include "fly/parser/exceptions.h"
-#include "fly/types/string.h"
+#include "fly/types/string/string.h"
 
 #include <cctype>
 #include <cstdio>
