+++ fly/system/nix/system_monitor_impl.cpp
@@ -3,7 +3,7 @@
 #include "fly/logger/logger.h"
 #include "fly/system/system_config.h"
 #include "fly/task/task_runner.h"
-#include "fly/types/string.h"
+#include "fly/types/string/string.h"
 
 #include <sys/sysinfo.h>
 #include <sys/times.h>
