+++ test/config/config_manager.cpp
@@ -4,7 +4,7 @@
 #include "fly/literals.h"
 #include "fly/path/path_config.h"
 #include "fly/task/task_manager.h"
-#include "fly/types/string.h"
+#include "fly/types/string/string.h"
 #include "test/config/test_config.h"
 #include "test/util/path_util.h"
 #include "test/util/waitable_task_runner.h"
