+++ test/config/test_config.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "fly/config/config.h"
-#include "fly/types/json.h"
+#include "fly/types/json/json.h"
 
 #include <string>
