+++ test/parser/json_parser.cpp
@@ -1,8 +1,8 @@
 #include "fly/parser/json_parser.h"
 
 #include "fly/parser/exceptions.h"
-#include "fly/types/json.h"
-#include "fly/types/string.h"
+#include "fly/types/json/json.h"
+#include "fly/types/string/string.h"
 
 #include <gtest/gtest.h>
