+++ test/parser/parser.cpp
@@ -1,7 +1,7 @@
 #include "fly/parser/exceptions.h"
 #include "fly/parser/json_parser.h"
-#include "fly/types/json.h"
-#include "fly/types/string.h"
+#include "fly/types/json/json.h"
+#include "fly/types/string/string.h"
 
 #include <gtest/gtest.h>
