+++ test/path/path_monitor.cpp
@@ -3,8 +3,8 @@
 #include "fly/literals.h"
 #include "fly/path/path_config.h"
 #include "fly/task/task_manager.h"
-#include "fly/types/concurrent_queue.h"
-#include "fly/types/string.h"
+#include "fly/types/concurrency/concurrent_queue.h"
+#include "fly/types/string/string.h"
 
 #include <gtest/gtest.h>
