+++ test/types/string_traits.cpp
@@ -1,7 +1,7 @@
-#include "fly/types/string_traits.h"
+#include "fly/types/string/string_traits.h"
 
 #include "fly/traits/traits.h"
-#include "fly/types/string_literal.h"
+#include "fly/types/string/string_literal.h"
 
 #include <gtest/gtest.h>
